+++ src/lib/Warp.NET.RenderImpl.Ui/Components/ScrollArea.cs
@@ -1,4 +1,5 @@
 using Warp.NET.Numerics;
+using Warp.NET.RenderImpl.Ui.Components.Styles;
 using Warp.NET.RenderImpl.Ui.Rendering;
 using Warp.NET.RenderImpl.Ui.Rendering.Scissors;
 using Warp.NET.Ui;
@@ -8,10 +9,13 @@ namespace Warp.NET.RenderImpl.Ui.Components;
 
 public class ScrollArea : AbstractScrollArea
 {
-	public ScrollArea(IBounds bounds, int scrollAmountInPixels, int scrollbarWidth)
+	public ScrollArea(IBounds bounds, int scrollAmountInPixels, int scrollbarWidth, ScrollAreaStyle scrollAreaStyle)
 		: base(bounds, scrollAmountInPixels, scrollbarWidth)
 	{
+		ScrollAreaStyle = scrollAreaStyle;
 	}
+
+	public ScrollAreaStyle ScrollAreaStyle { get; set; }
 
 	public override void Render(Vector2i<int> scrollOffset)
 	{
@@ -23,15 +27,15 @@ public class ScrollArea : AbstractScrollArea
 		ScissorScheduler.UnsetScissor();
 
 		// Render scrollbar.
-		Vector2i<int> borderVec = new(4);
+		Vector2i<int> borderVec = new(ScrollAreaStyle.BorderSize * 2);
 		Vector2i<int> scale = ScrollbarBounds.Size;
 
-		RenderImplUiBase.Game.RectangleRenderer.Schedule(scale, scrollOffset + ScrollbarBounds.Center, Depth, Color.White);
-		RenderImplUiBase.Game.RectangleRenderer.Schedule(scale - borderVec, scrollOffset + ScrollbarBounds.Center, Depth + 1, IsDraggingScrollbar ? Color.Gray(0.5f) : IsScrollbarHovering ? Color.Gray(0.25f) : Color.Black);
+		RenderImplUiBase.Game.RectangleRenderer.Schedule(scale, scrollOffset + ScrollbarBounds.Center, Depth, ScrollAreaStyle.BorderColor);
+		RenderImplUiBase.Game.RectangleRenderer.Schedule(scale - borderVec, scrollOffset + ScrollbarBounds.Center, Depth + 1, IsDraggingScrollbar ? ScrollAreaStyle.DraggingBackgroundColor : IsScrollbarHovering ? ScrollAreaStyle.HoveringBackgroundColor : ScrollAreaStyle.BackgroundColor);
 
-		const int thumbPadding = 8;
+		int thumbPadding = (ScrollAreaStyle.BorderSize + ScrollAreaStyle.ThumbPadding) * 2;
 		Vector2i<int> thumbScale = new(scale.X - thumbPadding, ScrollbarHeight - thumbPadding + 1); // + 1 is necessary for some reason.
 		Vector2i<int> thumbOffset = new(0, ScrollbarStartY);
-		RenderImplUiBase.Game.RectangleRenderer.Schedule(thumbScale, scrollOffset + new Vector2i<int>(ScrollbarBounds.Center.X, ScrollbarBounds.TopLeft.Y + thumbScale.Y / 2 + thumbPadding / 2) + thumbOffset, Depth + 2, Color.Gray(0.75f));
+		RenderImplUiBase.Game.RectangleRenderer.Schedule(thumbScale, scrollOffset + new Vector2i<int>(ScrollbarBounds.Center.X, ScrollbarBounds.TopLeft.Y + thumbScale.Y / 2 + thumbPadding / 2) + thumbOffset, Depth + 2, ScrollAreaStyle.ThumbColor);
 	}
 }