+++ src/lib/Warp.NET.RenderImpl.Ui/Components/Styles/ScrollAreaStyle.cs
@@ -1,0 +1,5 @@
+using Warp.NET.Numerics;
+
+namespace Warp.NET.RenderImpl.Ui.Components.Styles;
+
+public record ScrollAreaStyle(int BorderSize, int ThumbPadding, Color BorderColor, Color ThumbColor, Color BackgroundColor, Color DraggingBackgroundColor, Color HoveringBackgroundColor);